requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
python-multipart==0.0.6
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-socketio==5.9.0
websockets==11.0.3
SQLAlchemy==2.0.23
python-dotenv==1.0.0
Pillow==10.1.0
PyMySQL==1.1.0
email-validator==2.3.0
